requirements.txt
azure-core==1.23.0
azure-functions==1.9.0
azure-storage-blob==12.9.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
future==0.18.2
idna==3.3
isodate==0.6.1
msrest==0.6.21
numpy==1.22.2
oauthlib==3.2.0
pandas==1.4.1
pycparser==2.21
python-dateutil==2.8.2
python-twitter==3.5
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
typing_extensions==4.1.1
urllib3==1.26.8
